requirements.txt
# Minimal, stable dependencies required for the project.
Flask==3.0.3
APScheduler==3.10.4
requests==2.32.3
python-dotenv==1.0.1
